+++ s127/admin.py
@@ -1,11 +1,10 @@
-import copy
-
 import nested_admin
 from django.contrib import admin
 
 import s127.models
 from carting.admin import (
     GISModelAdminWithRasterMarine,
+    ModelAdminWithFormsetsIncludingInline,
     ModelAdminWithOrderedFormsets,
     Toto,
 )
@@ -348,9 +347,9 @@ class PilotServiceAdmin(ReportableServiceAreaAdmin):
 
 @admin.register(s127.models.FullPilotServiceProxy)
 class FullPilotServiceAdmin(
+    ModelAdminWithFormsetsIncludingInline,
     PilotServiceAdmin,
 ):
-    change_form_template = "admin/change_form_with_ordered_formsets_test.html"
     autocomplete_fields = ["pilotage_district"]
 
     list_display = (
@@ -359,34 +358,6 @@ class FullPilotServiceAdmin(
         "pilot_boarding_places",
     )
     list_filter = ("pilotage_district",)
-
-    def get_fieldsets(self, request, obj=None):
-        fieldsets = copy.deepcopy(self.fieldsets_and_inlines_ordered)
-        for fieldset in fieldsets:
-            fieldset[1]["fields"] = [
-                field for field in fieldset[1]["fields"] if isinstance(field, str)
-            ]
-        return fieldsets
-
-    def get_inlines(self, request, obj=None):
-        inlines = []
-        for fieldset in self.fieldsets_and_inlines_ordered:
-            inlines = inlines + [
-                field
-                for field in fieldset[1]["fields"]
-                if not isinstance(
-                    field, str
-                )  # FIXME: stronger to test if it is class inherit fro AdminInLine ?
-            ]
-        return inlines
-
-    def render_change_form(self, request, context, *args, **kwargs):
-        context.update(
-            {
-                "fieldsets_and_inlines_ordered": self.fieldsets_and_inlines_ordered,
-            }
-        )
-        return super().render_change_form(request, context, *args, **kwargs)
 
     fieldsets_and_inlines_ordered = [
         (
@@ -400,7 +371,7 @@ class FullPilotServiceAdmin(
         ("Pilotage districts", {"fields": ["pilotage_district"]}),
         (
             "Pilot Boarding Places",
-            {"fields": [PilotServicePilotBoardingPlaceInline, "pilotage_district"]},
+            {"fields": [PilotServicePilotBoardingPlaceInline]},
         ),
         ("Contact Details", {"fields": [SrvContactInline]}),
         ("Permission type", {"fields": [FeatureTypePermissionTypeInline]}),
