+++ static/to_compile/entrypoints/admin.css
@@ -58,3 +58,11 @@ label.required::after {
     z-index: 1;
     padding: 1rem 0;
 }
+
+fieldset fieldset h2 {
+    display: none;
+}
+
+fieldset .inline-group {
+    margin-bottom: 0px;
+}
